Worked out a Vulkan implementation of mouse-picking via integer framebuffer attachment (i.e. 'colour picking'). Will eventually compare this performance-wise to CPU-side raycast picking, but for now this will do

+++ Zahra/src/Platform/Vulkan/VulkanFramebuffer.cpp
@@ -14,7 +14,7 @@ namespace Zahra
 		if (m_Specification.Height == 0)
 			m_Specification.Height = Renderer::GetSwapchainHeight();
 
-		m_ColourAttachmentCount = m_Specification.ColourAttachmentSpecs.size();
+		m_ColourAttachmentCount = (uint32_t)m_Specification.ColourAttachmentSpecs.size();
 		CreateAttachments();
 	}
